sophia_api
Module dataset

An RDF dataset is composed of a default dataset, and zero or more named graphs, each associated with a dataset name.

Another way to look at it is as a collection of quads.

This module provides reusable abstractions for different kinds of datasets, as well as a few implementations for them.

Modules

adapter

I define adapters for the dataset related traits.

Traits

CollectibleDataset

A Dataset that can be constructed from a QuadSource

Dataset

Generic trait for RDF datasets.

MutableDataset

Generic trait for mutable RDF datasets.

SetDataset

Marker trait constraining the semantics of Dataset and MutableDataset.

Type Aliases

DQuadSource

Type alias for fallible quad iterators produced by a dataset.

DResult

Type alias for results produced by a dataset.

DTerm

Type alias for terms produced by a dataset.

DTermSource

Type alias for fallible term iterators produced by a dataset.

MdResult

Type alias for results produced by a mutable dataset.