sophia_api

Module graph

An RDF graph, the central notion of the RDF data model, is a collection of triples.

This module provides reusable abstractions for different kinds of graph, as well as a few implementations for them.

Modules

adapter

I define adapters for the graph related traits.

Traits

CollectibleGraph

A Graph that can be constructed from a TripleSource

Graph

Generic trait for RDF graphs.

MutableGraph

Generic trait for mutable RDF graphs.

SetGraph

Marker trait constraining the semantics of Graph and MutableGraph.

Type Aliases

GResult

Type alias for results produced by a graph.

GTerm

Type alias for terms produced by a graph.

GTermSource

Type alias for fallible term iterators produced by a graph.

GTripleSource

Type alias for fallible triple iterators produced by a graph.

MgResult

Type alias for results produced by a mutable graph.