sophia_api::graph::adapter

Struct PartialUnionGraph

pub struct PartialUnionGraph<D: Dataset, M: GraphNameMatcher> { /* private fields */ }

I wrap a Dataset as a Graph corresponding to the union of a subset of its graphs, determined by a GraphNameMatcher.

Implementations

impl<D: Dataset, M: GraphNameMatcher + Copy> PartialUnionGraph<D, M>

pub fn new(d: D, m: M) -> Self

Wrap the given dataset as a single Graph (selected via the given graph name).

pub fn unwrap(self) -> (D, M)

Unwrap the inner GraphName and Dataset.