sophia_api::graph

Trait MutableGraph

pub trait MutableGraph: Graph {
    type MutationError: Error + 'static;

    // Required methods
    fn insert<TS, TP, TO>(
        &mut self,
        s: TS,
        p: TP,
        p: TO,
    ) -> MgResult<Self, bool>
       where TS: Term,
             TP: Term,
             TO: Term;
    fn remove<TS, TP, TO>(
        &mut self,
        s: TS,
        p: TP,
        o: TO,
    ) -> MgResult<Self, bool>
       where TS: Term,
             TP: Term,
             TO: Term;

    // Provided methods
    fn insert_triple<T>(&mut self, triple: T) -> MgResult<Self, bool>
       where T: Triple { ... }
    fn remove_triple<T>(&mut self, triple: T) -> MgResult<Self, bool>
       where T: Triple { ... }
    fn insert_all<TS: TripleSource>(
        &mut self,
        src: TS,
    ) -> StreamResult<usize, TS::Error, <Self as MutableGraph>::MutationError> { ... }
    fn remove_all<TS: TripleSource>(
        &mut self,
        src: TS,
    ) -> StreamResult<usize, TS::Error, <Self as MutableGraph>::MutationError> { ... }
    fn remove_matching<S, P, O>(
        &mut self,
        ms: S,
        mp: P,
        mo: O,
    ) -> Result<usize, Self::MutationError>
       where S: TermMatcher,
             P: TermMatcher,
             O: TermMatcher,
             Self::MutationError: From<Self::Error> { ... }
    fn retain_matching<S, P, O>(
        &mut self,
        ms: S,
        mp: P,
        mo: O,
    ) -> Result<(), Self::MutationError>
       where S: TermMatcher,
             P: TermMatcher,
             O: TermMatcher,
             Self::MutationError: From<Self::Error> { ... }
}

Generic trait for mutable RDF graphs.

NB: the semantics of this trait allows a graph to contain duplicate triples; see also SetGraph.

Required Associated Types

type MutationError: Error + 'static

The error type that this graph may raise during mutations.

Required Methods

fn insert<TS, TP, TO>(&mut self, s: TS, p: TP, p: TO) -> MgResult<Self, bool>
where TS: Term, TP: Term, TO: Term,

Insert in this graph a triple made of the the given terms.

Return value

The bool value returned in case of success is not significant unless this graph also implements SetGraph.

If it does, true is returned iff the insertion actually changed the graph. In other words, a return value of false means that the graph was not changed, because the triple was already present in this SetGraph.

See also insert_triple

Usage

let schema = Namespace::new("http://schema.org/").unwrap();
let s_name = schema.get("name").unwrap();

graph.insert(&s_name, &rdf::type_, &rdf::Property)?;
graph.insert(&s_name, &rdfs::range, &xsd::string)?;
graph.insert(&s_name, &rdfs::comment, "The name of the item.")?;

fn remove<TS, TP, TO>(&mut self, s: TS, p: TP, o: TO) -> MgResult<Self, bool>
where TS: Term, TP: Term, TO: Term,

Remove from this graph the triple made of the the given terms.

Return value

The bool value returned in case of success is not significant unless this graph also implements SetGraph.

If it does, true is returned iff the removal actually changed the graph. In other words, a return value of false means that the graph was not changed, because the triple was already absent from this SetGraph.

Provided Methods

fn insert_triple<T>(&mut self, triple: T) -> MgResult<Self, bool>
where T: Triple,

Insert in this graph the given triple.

NB: if you want to insert a triple t while keeping its ownership, you can still pass t.spo().

See also MutableGraph::insert

fn remove_triple<T>(&mut self, triple: T) -> MgResult<Self, bool>
where T: Triple,

Remoe from this graph the given triple.

NB: if you want to remove a triple t while keeping its ownership, you can still pass t.spo().

See also MutableGraph::remove

fn insert_all<TS: TripleSource>( &mut self, src: TS, ) -> StreamResult<usize, TS::Error, <Self as MutableGraph>::MutationError>

Insert into this graph all triples from the given source.

Blank node scope

The blank nodes contained in the triple source will be inserted as is. If they happen to have the same identifier as blank nodes already present, they will be considered equal. This might not be what you want, especially if the graph contains data from a file, and you are inserting data from a different file. In that case, you should first transform the triple source, in order to get fresh blank node identifiers.

Return value

The usize value returned in case of success is not significant unless this graph also implements SetGraph.

If it does, the number of triples that were actually inserted (i.e. that were not already present in this SetGraph) is returned.

fn remove_all<TS: TripleSource>( &mut self, src: TS, ) -> StreamResult<usize, TS::Error, <Self as MutableGraph>::MutationError>

Remove from this graph all triples from the given source.

Return value

The usize value returned in case of success is not significant unless this graph also implements SetGraph.

If it does, the number of triples that were actually removed (i.e. that were not already absent from this SetGraph) is returned.

fn remove_matching<S, P, O>( &mut self, ms: S, mp: P, mo: O, ) -> Result<usize, Self::MutationError>
where S: TermMatcher, P: TermMatcher, O: TermMatcher, Self::MutationError: From<Self::Error>,

Remove all triples matching the given matchers.

Return value

The usize value returned in case of success is not significant unless this graph also implements SetGraph.

If it does, the number of triples that were actually removed (i.e. that were not already absent from this SetGraph) is returned.

Note to implementors

The default implementation is rather naive, and could be improved in specific implementations of the trait.

fn retain_matching<S, P, O>( &mut self, ms: S, mp: P, mo: O, ) -> Result<(), Self::MutationError>
where S: TermMatcher, P: TermMatcher, O: TermMatcher, Self::MutationError: From<Self::Error>,

Keep only the triples matching the given matchers.

Note to implementors

The default implementation is rather naive, and could be improved in specific implementations of the trait.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.