sophia_api::graph

Trait SetGraph

pub trait SetGraph: Graph { }

Marker trait constraining the semantics of Graph and MutableGraph.

It guarantees that (1) triples will never be returned / stored multiple times.

If the type also implements MutableGraph, it must also ensure that (2) the bool or usize values returned by MutableGraph methods accurately describe how many triples were actually added/removed.

Note to implementors

A type implementing both Graph and MutableGraph, enforcing (1) but failing to enforce (2) must not implement this trait.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.