sophia_api::ns

Struct Namespace

pub struct Namespace<T: Borrow<str>>(/* private fields */);

A custom namespace.

The get method can be used to create a new IRI by concatenating a suffix to this namespace’s IRI.

Implementations

impl<T: Borrow<str>> Namespace<T>

pub fn new(iri: T) -> Result<Self, InvalidIri>

Build a custom namespace based on the given IRI.

iri must be a valid IRI, otherwise this constructor returns an error.

pub fn new_unchecked(iri: T) -> Self

Build a custom namespace based on the given IRI, without checking that it is valid. If it is not, it may result in undefined behaviour.

pub fn get<'s>(&'s self, suffix: &'s str) -> Result<NsTerm<'s>, InvalidIri>

Build an IRI by appending suffix to this namespace.

Return an error if the concatenation produces an invalid IRI.

pub fn get_unchecked<'s>(&'s self, suffix: &'s str) -> NsTerm<'s>

Build an IRI by appending suffix to this namespace, without checking that the resulting IRI is valid. If it is not, it may result in undefined behaviour.

pub fn inner(self) -> IriRef<T>

Consume this Namespace and return the inner IriRef.

impl Namespace<&'static str>

pub const fn new_unchecked_const(iri: &'static str) -> Self

const constructor for Namespace

Precondition

iri must be a valid IRI reference (possibly relative), otherwise undefined behaviour may occur.