sophia_api::ns

Struct NsTerm

pub struct NsTerm<'a> { /* private fields */ }

A Term produced by a Namespace.

The raison d’être of this type, compared to IriRef<&str>, is that it stored the IRI in two parts (namespace and suffix), so that the namespace can be reused by multiple distinct terms.

It makes sense for Namespaces, whose terms all have the same prefix.

Implementations

impl<'a> NsTerm<'a>

pub const fn new_unchecked(ns: IriRef<&'a str>, suffix: &'a str) -> Self

Make an NsTerm without checking that it produces a valid IRI.

pub fn iriref(&self) -> IriRef<MownStr<'_>>

Return an IriRef representing this term.

pub fn to_iriref(self) -> IriRef<MownStr<'a>>

Return an IriRef representing this term.