sophia_api

Module serializer

API for serializing RDF syntaxes.

This module specifies, through dedicated traits, how serialization of triples/graphs and quads/dataset is handled in sophia. These traits are designed with the idea that each serializer has a specific “target” (typically a file or something similar) associated to it. If you want to serialize to two different files, you must create two different serializers.

Note however that this API does not cover the creation of serializers, and therefore does not cover how their target is specified.

Traits

QuadSerializer

A quad serializer writes quads according to a given format.

Stringifier

A stringifier is special kind of TripleSerializer or QuadSerializer:

TripleSerializer

A triple serializer writes triples according to a given format.