sophia_api::source

Type Alias QSQuad

pub type QSQuad<'a, TS> = <TS as IsQuadSource>::Quad<'a>;

Type alias to denote the type of quads yielded by a QuadSource.

Why not using TS::Item<'a> instead? Source::Item being a generic associated type (GAT), the compiler will not always “know” that TS::Item<'a> implements the Quad trait. This type alias, on the other hand, will always be recognized as a Quad implementation.