sophia_api::term::bnode_id

Struct BnodeId

pub struct BnodeId<T: Borrow<str>>(/* private fields */);

See BnodeId::new.

Implementations

impl<T: Borrow<str>> BnodeId<T>

pub fn new(id: T) -> Result<Self, InvalidBnodeId>

This wrapper guarantees that the underlying str satisfies the BLANK_NODE_LABEL rule in Turtle (without the leading _:).

pub fn as_str(&self) -> &str

Gets a reference to the underlying &str.

pub fn new_unchecked(inner: T) -> Self

Construct a BnodeId<T> without checking that the inner value is valid. If it is not, it may result in undefined behaviour.

pub fn unwrap(self) -> T

Returns the wrapped value, consuming self.

pub fn map_unchecked<F, U>(self, f: F) -> BnodeId<U>
where F: FnOnce(T) -> U, U: Borrow<str>,

Map a BnodeId<T> to a BnodeId<U> by applying a function to the wrapped value. It does not check that the value returned by the function is valid. If it is not, it may result in undefined behaviour.

impl<T> BnodeId<T>
where T: Borrow<str>,

pub fn as_ref(&self) -> BnodeId<&str>

Convert from &BnodeId<T> to BnodeId<&str>.

impl BnodeId<&'static str>

pub const fn new_unchecked_const(inner: &'static str) -> Self

Construct a BnodeId<&'static str> without checking that the inner value is valid. If it is not, it may result in undefined behaviour.