sophia_api::term

Enum SimpleTerm

pub enum SimpleTerm<'a> {
    Iri(IriRef<MownStr<'a>>),
    BlankNode(BnodeId<MownStr<'a>>),
    LiteralDatatype(MownStr<'a>, IriRef<MownStr<'a>>),
    LiteralLanguage(MownStr<'a>, LanguageTag<MownStr<'a>>),
    Triple(Box<[Self; 3]>),
    Variable(VarName<MownStr<'a>>),
}

A straighforward implementation of Term as an enum.

Variants

Iri(IriRef<MownStr<'a>>)

An RDF IRI

BlankNode(BnodeId<MownStr<'a>>)

An RDF blank node

LiteralDatatype(MownStr<'a>, IriRef<MownStr<'a>>)

An RDF literal

LiteralLanguage(MownStr<'a>, LanguageTag<MownStr<'a>>)

An RDF language-tagged string

Triple(Box<[Self; 3]>)

An RDF-star quoted triple

Variable(VarName<MownStr<'a>>)

A SPARQL or Notation3 variable

Implementations

impl<'a> SimpleTerm<'a>

pub fn from_term_ref<T>(term: &'a T) -> Self
where T: Term + ?Sized,

Build a SimpleTerm that borrows as much as possible from the original term.

NB: depending on the implementation of term, some data might still be allocated.

pub fn from_triple<T: Triple>(triple: T) -> Self

Build a SimpleTerm of kind Triple from any triple.