sophia_api::term

Module matcher

I define generic traits and default implementations for matchers, objects that can be used to match zero, one or several terms.

For a list of matcher implementations, check TermMarcher’s and GraphNameMatcher’s implementors lists.

For methods using matchers (with examples), see for example Triple::matched_by, Graph::triples_matching, MutableGraph::remove_matching, MutableGraph::retain_matching, Dataset::quads_matching, MutableDataset::remove_matching, MutableDataset::retain_matching.

Structs

Any

A universal matcher: it matches any Term or GraphName (even the default graph).

DatatypeMatcher

A TermMatcher that matches all literals with a given datatype

LanguageTagMatcher

A TermMatcher that matches all literals with a given language tag.

MatcherRef

Result type of TermMatcher::matcher_ref and GraphNameMatcher::matcher_ref.

TermMatcherGn

Wrapper type returned by TermMatcher::gn

Traits

GraphNameMatcher

Generic trait for matching GraphNames.

TermMatcher

Generic trait for matching Terms.