sophia_api::term::matcher

Trait GraphNameMatcher

pub trait GraphNameMatcher {
    type Term: Term + ?Sized;

    // Required method
    fn matches<T2: Term + ?Sized>(&self, graph_name: GraphName<&T2>) -> bool;

    // Provided methods
    fn constant(&self) -> Option<GraphName<&Self::Term>> { ... }
    fn matcher_ref(&self) -> MatcherRef<'_, Self> { ... }
}

Generic trait for matching GraphNames.

Required Associated Types

type Term: Term + ?Sized

The type of term that this GraphNameMatcher contains

Required Methods

fn matches<T2: Term + ?Sized>(&self, graph_name: GraphName<&T2>) -> bool

Check whether this matcher matches t.

Provided Methods

fn constant(&self) -> Option<GraphName<&Self::Term>>

Return None, unless this matcher can only match a single graph name, in which case this method may return that graph name.

This method is provided for optimization purposes, so implementing it is optional.

fn matcher_ref(&self) -> MatcherRef<'_, Self>

Return a GraphNameMatcher that is actually just a reference to this one.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.