sophia_api::term::matcher
Trait TermMatcher

pub trait TermMatcher {
    type Term: Term + ?Sized;

    // Required method
    fn matches<T2: Term + ?Sized>(&self, term: &T2) -> bool;

    // Provided methods
    fn constant(&self) -> Option<&Self::Term> { ... }
    fn gn(self) -> TermMatcherGn<Self>
       where Self: Sized { ... }
    fn matcher_ref(&self) -> MatcherRef<'_, Self> { ... }
}

Generic trait for matching Terms.

Required Associated Types

type Term: Term + ?Sized

The type of term that this TermMatcher contains

Required Methods

fn matches<T2: Term + ?Sized>(&self, term: &T2) -> bool

Check whether this matcher matches t.

Provided Methods

fn constant(&self) -> Option<&Self::Term>

Return None, unless this matcher can only match a single term, in which case this method may return that term.

This method is provided for optimization purposes, so implementing it is optional.

fn gn(self) -> TermMatcherGn<Self>
where Self: Sized,

Converts this TermMatcher into a GraphNameMatcher

If you only want to borrow this matcher as a GraphNameMatcher, call gn on the result of matcher_ref.

fn matcher_ref(&self) -> MatcherRef<'_, Self>

Return a TermMatcher that is actually just a reference to this one.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.