sophia_api::term

Trait TryFromTerm

pub trait TryFromTerm: Sized {
    type Error: 'static + Error;

    // Required method
    fn try_from_term<T: Term>(term: T) -> Result<Self, Self::Error>;
}

A type that can be built from some terms.

See also FromTerm

Required Associated Types

type Error: 'static + Error

The error type produced when failing to copy a given term

Required Methods

fn try_from_term<T: Term>(term: T) -> Result<Self, Self::Error>

Try to copy term into an instance of this type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.